fix-gotoolbox-parent-json-file (#228)
* feat: Exclude additional directories from dagger build

Exclude the following directories from the Dagger build:
- .vscode
- .idea
- .trunk
This ensures that development-specific files and directories are not included in the build, improving the overall build process.

* feat: Upgrade Dagger engine version to v0.13.5

This commit updates the Dagger engine version from v0.13.3 to v0.13.5 in the "module-template-light/dagger.json" and "gotoolbox/dagger.json" files. This upgrade will provide the latest features and bug fixes from the Dagger engine.