fix: Change the "keepGitDir" parameter to "discardGitDir" in the Git clone function (#229)
* fix: Change the "keepGitDir" parameter to "discardGitDir" in the Git clone function

The changes in this commit update the "WithClonedGitRepoSSH" and "CloneGitRepoSSH" functions to use the new "discardGitDir" parameter instead of the deprecated "keepGitDir" parameter. This change was introduced in version 0.13.4 of the module-template-light library.

* feat: Update Dagger versions in module-template-light CI workflow

Updates the Dagger versions used in the module-template-light CI workflow from 0.13.0, 0.13.1, 0.13.2, 0.13.3 to 0.13.4 and 0.13.5. This change ensures the workflow runs tests and recipes against the latest Dagger versions.