terragrunt-module-adjustments (#232)
* feat: Improve container environment variable validation

- Rename `utilValidateIfEnvVarIsSetInContainer` to `assertEnvVarIsSetInContainer` for better clarity
- Simplify the implementation of `assertEnvVarIsSetInContainer` by using `test -d` instead of `ls -la`
- Rename `utilTheseFilesExistsInContainer` to `assertTheseFilesExistsInContainer` for better clarity
- Simplify the implementation of `assertTheseFilesExistsInContainer` by using `test -f` instead of `ls -la`
- Rename `utilFileShouldContainContent` to `assertFileContentShouldContain` for better clarity

* fix: Remove unnecessary assertion in `TestTerragruntExecLifecycleCommands`

The changes in this commit remove the unnecessary assertion that checks the output of the `terraform plan` command in the `TestTerragruntExecLifecycleCommands` test. This assertion was redundant as the test already asserts that the command executed successfully, and the plan output is not the focus of this particular test.

Additionally, the changes improve the `assertFileContentShouldContain` function by simplifying the logic and removing unused code. The function now directly checks if the file exists and if its content contains the expected string, without unnecessary intermediate steps.

* feat: Add Terragrunt apply and destroy commands with auto-approve

This commit adds the implementation of the Terragrunt apply and destroy commands with the auto-approve flag. The changes are as follows:

- Add a new block of code to execute the `apply` command with the `-auto-approve` flag as an argument.
- Add a new block of code to execute the `destroy` command with the `AutoApprove` option set to `true`.
- These changes ensure that the Terragrunt commands are executed successfully without requiring manual approval.