feat: Add test for Terragrunt exec with plan output
This commit adds a new test case to the `TestAll` function in the `main.go` file of the `terragrunt/tests` package. The new test case, `TestTerragruntExecWithPlanOutput`, verifies the behavior of the Terragrunt `exec` command when it is executed with the `--plan-output-path` option. This change enhances the test suite to cover more use cases for the Terragrunt `exec` command.