feat: Add recipe to format Go files in Dagger modules (#234)
The changes introduced in this commit add a new recipe called `fmt` to the `justfile` that formats Go files in Dagger modules, excluding the `internal/` directory and `dagger.gen.go` files. The recipe checks and formats the Go files in the parent module and its submodules (`tests` and `examples/go`), if they exist. It also provides feedback on the formatted files and skips non-Dagger modules.