add-support-to-disabling-providers-cache (#241)
* feat: Disable Terragrunt provider cache server

Adds a new function `WithTerragruntProviderCacheServerDisabled` to the `Terragrunt` struct. This function sets the `TERRAGRUNT_PROVIDER_CACHE` environment variable to "0" to disable the Terragrunt provider cache server. This can be useful in some cases where you may want to disable the cache server.

* feat: Add function to configure extra registries for Terragrunt provider cache

This change adds a new function `WithRegistriesToCacheProvidersFrom` to the `Terragrunt` struct in the `terragrunt_cfg.go` file. This function allows users to specify additional registries to cache providers from, beyond the default registries of `registry.terraform.io` and `registry.opentofu.org`.

The function takes a slice of strings representing the registries to cache providers from, and appends them to the default list. It then sets the `TERRAGRUNT_PROVIDER_CACHE_REGISTRY_NAMES` environment variable with the comma-separated list of registries.

This feature enhances the flexibility of the Terragrunt provider cache configuration, allowing users to cache providers from custom registries if needed.

* feat: Add support for custom provider cache registries and disable provider cache server

This commit introduces the following changes:

1. Add support for configuring custom provider cache registries in Terragrunt using the `WithRegistriesToCacheProvidersFrom` method.
   - The `TestTerragruntWithCustomRegistriesToCacheProvidersFrom` function ensures that the `TERRAGRUNT_PROVIDER_CACHE_REGISTRY_NAMES` environment variable is set correctly.

2. Add support for disabling the provider cache server in Terragrunt using the `WithTerragruntProviderCacheServerDisabled` method.
   - The `TestTerragruntWithProviderCacheServerDisabled` function ensures that the `TERRAGRUNT_PROVIDER_CACHE` environment variable is set to "0".

3. Update the `TestAll` function in the `main.go` file to include the new test cases for custom provider cache registries and the provider cache server.

These changes improve the flexibility and configurability of the Terragrunt module, allowing users to customize the provider caching behavior to better suit their needs.