feat-add-native-cmd-support (#243)
* feat: Add Entrypoint and IsValidIACTool functions to config.go

This commit introduces the Entrypoint type and the IsValidIACTool function in the config.go file. The Entrypoint type represents the entrypoint to use when executing the command, with three possible values: TerragruntEntrypoint, TerraformEntrypoint, and OpentofuEntrypoint. The IsValidIACTool function validates the provided IaC tool and returns an error if the tool is not one of the valid values.

These changes are part of the ongoing effort to improve the flexibility and robustness of the tool's configuration handling.

* feat: Expand the Cmd interface to support more IaC commands

The changes to the `iac_cmd.go` file include:

1. Expanding the `Cmd` interface to support additional functionality:
   - `ExecCmd`: Executes a given command within a Dagger container.
   - `TfExecCmd`: Executes a given Terraform command within a Dagger container.
   - `TgExecCmd`: Executes a given Terragrunt command within a Dagger container.
2. Adding detailed documentation for the new methods, including parameter descriptions and return values.
3. Improving the existing `Exec` method by adding a new `entrypoint` parameter to allow for more flexibility in executing IaC commands.
4. Adding validation for the provided command in the `validate` method to ensure it is recognized by the IaC tool.

These changes aim to provide a more comprehensive and flexible interface for executing IaC commands within a Dagger container, supporting a wider range of use cases and improving the overall robustness of the codebase.

* feat: Add support for specifying IaC tool in Terragrunt execution

The main changes in this commit are:

1. Introduce a new `Tool` type to represent the IaC tool to be used for executing the command. This allows the user to specify the tool (Terragrunt, Terraform, or Opentofu) when executing a command.

2. Update the `Exec` and `ExecCmd` methods in the `Terragrunt` struct to accept an optional `tool` parameter. If the `tool` parameter is provided, it is used to set the entrypoint for the Dagger container. If not provided, the default entrypoint (Terragrunt) is used.

3. Update the `IsValidIACTool` function to validate the provided tool against the list of supported tools.

4. Rename the `TgExecCmd` method to `TfExecCmd` to better reflect its purpose of executing Terraform commands.

5. Remove the `Entrypoint` parameter from the `Exec` and `TfExecCmd` methods, as it is now replaced by the `tool` parameter.

This change simplifies the API and provides more flexibility for users to choose the IaC tool they want to use when executing commands within the Dagger container.

* feat: Add new Terragrunt test for simple 'terraform init' command

Implement a new test function `TestTfExecInitSimpleCommand` that sets up the necessary environment variables, initializes a Terragrunt module, and executes the 'init' command. It then validates the output of the command and checks if the environment variables are correctly set in the container. If any step fails, an error is returned.

Also, add new function documentation for `TfExecCmd` and `TfExec` methods in the `terragrunt_cmds.go` file to provide more clarity on their purpose and usage.

* feat: Improve terraform execution flow

- Update the TestTfExecInitSimpleCommand test to initialize the terraform module in the correct directory (terraform/tf-module-1)
- Add an additional ls command to the terraform plan execution to better understand the state of the module
- Refactor the plan execution to use the WithExec method instead of the TfExec method for more control over the command execution
- Remove the commented-out code block for the previous plan execution approach

* feat: Improve Terragrunt command execution and testing

v0.1.0 - Remove unused Terragrunt command execution methods
- Remove unused `TfExec` and `TfExecCmd` methods from the `Cmd` interface
- Simplify the `Exec` method in `terragrunt_cmds.go` by removing unnecessary code

v0.2.0 - Add support for executing multiple Terraform commands
- Add support for executing multiple Terraform commands in a single container
- Update the `TestTfExecInitSimpleCommand` test to use the new `WithTerraformCommand` method
- Add the `TestTfExecInitSimpleCommand` test to the `TestAll` method in `main.go`

v0.3.0 - Update the base Alpine Docker image configuration
- Update the `base-alpine.yaml` file with the latest package versions
- Add the `terraformProvidersDir` to the directories with default permissions