feat: Remove unnecessary 'detailed' option from sync and inspect tasks
This change removes the 'detailed' option from the sync_modules_task and inspect_modules_task functions. The 'detailed' option was previously used to provide more detailed information about the changes, but it has been determined that this level of detail is not necessary for the sync and inspect tasks. By removing this option, the code becomes more concise and easier to maintain.