fix: Update test directory reference in TestWithConfigFile method
The changes in this commit update the reference to the test directory in the `TestWithConfigFile` method in both `module-template-light/tests/apis.go` and `module-template/tests/apis.go` files.

Previously, the test directory was accessed using the `m.getTestDir()` method, but this has been changed to use the `m.TestDir` property directly.

This change ensures that the test directory is correctly referenced and used in the `TestWithConfigFile` method.