feat: Update dagger and daggerx dependencies to latest versions
# Description
The changes in this commit update the `dagger.io/dagger` and `github.com/Excoriate/daggerx` dependencies to the latest versions `v0.13.0` and `v0.0.32`, respectively. This ensures that the project is using the most up-to-date versions of these dependencies, which may include bug fixes, performance improvements, or new features.

The `go.mod` and `go.sum` files have been updated to reflect these changes.