feat: Update dagger and daggerx dependencies to latest versions (#249)
* feat: Update dagger and daggerx dependencies to latest versions

# Description
The changes in this commit update the `dagger.io/dagger` and `github.com/Excoriate/daggerx` dependencies to the latest versions `v0.13.0` and `v0.0.32`, respectively. This ensures that the project is using the most up-to-date versions of these dependencies, which may include bug fixes, performance improvements, or new features.

The `go.mod` and `go.sum` files have been updated to reflect these changes.

* feat: Add expected return codes for container commands

docs: Improve documentation for Container.Stdout and Container.Stderr

The changes in this commit:

1. Add a new field `Expect` to the `ContainerWithExecOpts` struct, which allows specifying the expected return codes for a command executed in the container. This provides more flexibility in handling command execution.

2. Improve the documentation for the `Container.Stdout` and `Container.Stderr` methods to clarify that they will return an error if no command has been set, instead of executing a default command.

* feat: Update OpenTelemetry dependencies to v1.27.0

This commit updates the OpenTelemetry dependencies in the go.mod files for the following modules:
- module-template-light
- tflinter
- module-template
- gotoolbox
- terragrunt

The changes include updating the following packages to version 1.27.0:
- `go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc`
- `go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp`
- `go.opentelemetry.io/otel/metric`
- `go.opentelemetry.io/otel/sdk/metric`

Additionally, the `golang.org/x/sys` version is updated to `v0.26.0`.

* feat: Update dependencies for Go modules

The commit updates the Go module dependencies across multiple modules:

- Upgrade the `go.opentelemetry.io/otel/metric` and `go.opentelemetry.io/otel/sdk/metric` dependencies to version 1.27.0.
- Add new dependencies for `go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc` and `go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp` version 1.27.0.
- Update the `golang.org/x/sys` dependency to version 0.26.0.
- Upgrade the `github.com/99designs/gqlgen` dependency to version 0.17.55 across the module-template and module-template-light modules.

* feat: Add Alpine container with utilities (#266)

Add support for configuring an Alpine container with common utilities like curl.
This includes:
- A test to verify the presence of the 'curl' utility in the Alpine container.
- Support for configuring an Ubuntu container with utilities.
- A test to verify the presence of the 'curl' utility in the Ubuntu container.
- Support for configuring an Alpine container with Git.
- A test to verify the presence of Git in the Alpine container.

* feat: Update dagger and daggerx dependencies to latest versions

# Description
The changes in this commit update the `dagger.io/dagger` and `github.com/Excoriate/daggerx` dependencies to the latest versions `v0.13.0` and `v0.0.32`, respectively. This ensures that the project is using the most up-to-date versions of these dependencies, which may include bug fixes, performance improvements, or new features.

The `go.mod` and `go.sum` files have been updated to reflect these changes.