feat: Update Dagger engine and Docker versions
This commit updates the default versions of the Dagger engine and Docker engine used in the Dagger module and test code.

The changes include:
- Updating the default Dagger engine version from "v0.13.0" to "v0.13.7"
- Updating the Dagger engine version used in the test suite from "v0.13.0" to "v0.13.6" and "v0.13.7"
- Updating the default Docker engine version from "24.0" to "24.0"

These changes ensure that the module and tests are using the latest stable versions of the Dagger engine and Docker engine, which may include bug fixes and performance improvements.