aa-branch-3 (#269)
* feat: Implement container and environment variable setup functionality

This commit adds two new functions to the `ModuleTemplate` and `ModuleTemplateLight` structs: `setupContainer` and `setupEnvironmentVariables`.

The `setupContainer` function sets up the container based on whether a custom image is provided or not. If a custom image is provided, it validates the image URL and sets the base container. If no custom image is provided, it sets up the default image using the `getImageURL` function.

The `setupEnvironmentVariables` function sets up the environment variables for the container. If environment variables are provided, it converts them to a list of `DaggerEnvVars` and adds them to the container.

These changes improve the modularity and testability of the code by separating the container and environment variable setup logic into dedicated functions.

* feat: Add methods to install GCC compiler and musl-dev package for Go CGO support

Adds `WithGoGCCCompilerAlpine()` and `WithGoGCCCompilerUbuntu()` methods to the `ModuleTemplate` and `ModuleTemplateLight` structs. These methods install the GCC compiler and musl-dev package in the container environment, which is necessary for enabling the Go toolchain to compile C code and link against C libraries when using CGO.

* feat: Update Dagger version to 0.13.7

- Update the default Dagger version in `module-template/dagger.go` from 0.13.0 to 0.13.7
- Update the Dagger version used in the tests in `module-template/tests/dagger.go` from 0.13.3 to 0.13.7
- Update the Dagger version documentation in `module-template/dagger.go` to use 0.13.7 as the example version
- Update the Dagger version used in the `testDaggerVersion` function in `module-template/tests/dagger.go` from 0.13.0, 0.13.1, 0.13.2, 0.13.3 to 0.13.6, 0.13.7

* feat: Update Dagger engine and Docker versions

This commit updates the default versions of the Dagger engine and Docker engine used in the Dagger module and test code.

The changes include:
- Updating the default Dagger engine version from "v0.13.0" to "v0.13.7"
- Updating the Dagger engine version used in the test suite from "v0.13.0" to "v0.13.6" and "v0.13.7"
- Updating the default Docker engine version from "24.0" to "24.0"

These changes ensure that the module and tests are using the latest stable versions of the Dagger engine and Docker engine, which may include bug fixes and performance improvements.

* feat: Update module dependencies to the latest versions

Upgrade the Excoriate/daggerx dependency to version 0.0.33 in both the module-template and module-template-light projects. This ensures the latest features and bug fixes are available in the projects.