feat: Update Go version to 1.21 in tests (#274)
* feat: Update Go version to 1.21 in tests

The changes update the Go version used in the `TestGoWithGoTestSum` test from 1.20.4 to 1.21. This ensures the tests are running with the latest stable version of Go.

* feat: Update Go version to 1.21 in tests

The changes update the Go version used in the `TestGoWithGoTestSum` test from 1.20.4 to 1.21. This ensures the tests are running with the latest stable version of Go.