fix: Update Dagger version to v6 in CI workflows (#276)
The changes in this commit update the Dagger version used in the CI workflows from v7 to v6. This is done to address a compatibility issue with the latest version of Dagger.

The changes touch multiple files in the `.github/workflows/` directory, where the Dagger version is updated in the following places:

- `ci-mod-module-template.yaml`
- `ci-mod-module-template-light.yaml`

The commit also updates the `DAGGER_VERSION` environment variable to use a specific version (0.14.0) in the `ci-mod-module-template.yaml` file.

This change is necessary to ensure the CI workflows continue to function correctly with the current version of Dagger.