ci: Simplify job names in CI workflows
The changes in this commit simplify the job names in the CI workflows for the module templates. The job names no longer include the Dagger version, as this information is not necessary for the job descriptions.

The changes are as follows:

- In the `ci-mod-templates.yaml` file, the job names have been updated to remove the Dagger version information. For example, the job name "🏗️ Build ${{ matrix.module }} with Dagger ${{ env.DAGGER_VERSION }}" has been changed to "🏗️ Build ${{ matrix.module }}".
- Similar changes have been made to the job names for the "Run Recipes", "Run Tests", and "Call Functions" jobs.

This simplifies the job names and makes them more concise, while still providing the necessary information about the module being built or tested.