ci: Update CI workflow to list main module functions
The changes in this commit update the CI workflow to list the main module functions for each module in the repository. The `workdir` and `module` inputs in the `dagger-for-github` action have been updated to use the `${{ matrix.module }}` variable, which allows the action to be run for each module in the matrix. This simplifies the workflow and ensures that the main module functions are listed for all modules.