ci: Update Dagger and GolangCI-Lint versions
This commit updates the versions of Dagger and GolangCI-Lint used in the CI workflow. Specifically:

- Bumps the Dagger version to 0.14.0
- Bumps the GolangCI-Lint version to 1.59.1

The purpose of these updates is to ensure the CI workflow is using the latest stable versions of these tools, which can provide bug fixes, performance improvements, and new features.

Additionally, the commit adds some `ls -ltrah` commands to the workflow steps to help with debugging and visibility of the current working directory.