ci: Add CI workflow for aws-tag-inspector
This commit adds a new GitHub Actions workflow for the aws-tag-inspector
module. The workflow includes the following steps:

1. Lint the aws-tag-inspector module using Dagger and GolangCI-Lint.
2. Run tests for the aws-tag-inspector module using Dagger.
3. Run the examples/go module for aws-tag-inspector using Dagger.

The workflow is triggered on push and pull request events for the main and
master branches, as well as on manual workflow dispatch. The workflow
uses the latest version of Dagger (0.14.0) and Go 1.22.